+++ packages/sdk-next/test/script/examples/router/getBestBuy.ts
@@ -16,7 +16,7 @@ class GetBestBuy extends PapiExecutor {
     const utils = new sor.TradeUtils(client);
 
     const buy = await router.getBestBuy(10, 5, 100_000_000_000n);
-    const tx = await utils.buildTx(buy);
+    const tx = await utils.buildBuyTx(buy);
     console.log(buy.toHuman());
     console.log('Transaction hash: ' + tx.asHex());